Created module_id_t and error_t types
It's primitive, but a start. Each module in the system now must
define a 4 bit id (module_id_t). It can then define up to 16
errors unique to itself, where the upper 4-bits of the byte are
the error id, and the lower 4-bits are the module id.

When an error propegates up to the application, it will blink the
8 bit error code on the LEDs for a few seconds, then return.